Created an individual class for the Sensor Calibrations. Fixed the notebox so that the notes save with the configuration. (still needs to be tested, but it looks like everything works)
